changes for filepath show | update ui for dio logger
