+++ doc/api/__404error.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, height=device-height, initial-scale=1, user-scalable=no">
-  <meta name="generator" content="made with love by dartdoc 7.0.1-wip">
+  <meta name="generator" content="made with love by dartdoc 8.0.3">
   <meta name="description" content="snug_logger API docs, for the Dart programming language.">
   <title>snug_logger - Dart API docs</title>
 
@@ -93,7 +93,7 @@
 <footer>
   <span class="no-break">
     snug_logger
-      1.0.4
+      1.0.8
   </span>
