+++ doc/api/snug_logger/LogType-enum-sidebar.html
@@ -9,6 +9,7 @@
           <li><a href="snug_logger/LogType.html#production">production</a></li>
           <li><a href="snug_logger/LogType.html#error">error</a></li>
 
+    
         <li class="section-title inherited">
           <a href="snug_logger/LogType.html#instance-properties">Properties</a>
         </li>
@@ -23,7 +24,9 @@
         <li class="section-title inherited"><a href="snug_logger/LogType.html#operators">Operators</a></li>
           <li class="inherited"><a href="https://api.flutter.dev/flutter/dart-core/Object/operator_equals.html">operator ==</a></li>
 
+    
 
+    
 
 
         <li class="section-title"><a href="snug_logger/LogType.html#constants">Constants</a></li>
