+++ doc/api/snug_logger/LogType.html
@@ -29,7 +29,7 @@
   <span id="sidenav-left-toggle" class="material-symbols-outlined" role="button" tabindex="0">menu</span>
   <ol class="breadcrumbs gt-separated dark hidden-xs">
     <li><a href="../index.html">snug_logger</a></li>
-    <li><a href="../snug_logger/snug_logger-library.html">snug_logger</a></li>
+    <li><a href="../snug_logger/snug_logger-library.html">snug_logger.dart</a></li>
     <li class="self-crumb">LogType enum</li>
   </ol>
   <div class="self-name">LogType</div>
@@ -283,9 +283,6 @@ Warning: This log will be displayed in the release mode.</p>
   A constant List of the values in this enum, in order of their declaration.
   
 
-    <div>
-      <span class="signature"><code>[info, debug, production, error]</code></span>
-    </div>
 </dd>
 
     </dl>
@@ -319,7 +316,7 @@ Warning: This log will be displayed in the release mode.</p>
 <footer>
   <span class="no-break">
     snug_logger
-      1.0.4
+      1.0.8
   </span>
