+++ doc/api/snug_logger/LogType/values-constant.html
@@ -29,7 +29,7 @@
   <span id="sidenav-left-toggle" class="material-symbols-outlined" role="button" tabindex="0">menu</span>
   <ol class="breadcrumbs gt-separated dark hidden-xs">
     <li><a href="../../index.html">snug_logger</a></li>
-    <li><a href="../../snug_logger/snug_logger-library.html">snug_logger</a></li>
+    <li><a href="../../snug_logger/snug_logger-library.html">snug_logger.dart</a></li>
     <li><a href="../../snug_logger/LogType.html">LogType</a></li>
     <li class="self-crumb">values constant</li>
   </ol>
@@ -105,7 +105,7 @@
 <footer>
   <span class="no-break">
     snug_logger
-      1.0.4
+      1.0.8
   </span>
