+++ doc/api/snug_logger/SnugDioLogger-class-sidebar.html
@@ -4,6 +4,7 @@
           <li><a href="snug_logger/SnugDioLogger/SnugDioLogger.html">SnugDioLogger</a></li>
 
 
+    
         <li class="section-title">
           <a href="snug_logger/SnugDioLogger-class.html#instance-properties">Properties</a>
         </li>
@@ -26,7 +27,9 @@
         <li class="section-title inherited"><a href="snug_logger/SnugDioLogger-class.html#operators">Operators</a></li>
           <li class="inherited"><a href="https://api.flutter.dev/flutter/dart-core/Object/operator_equals.html">operator ==</a></li>
 
+    
 
+    
 
 
 </ol>
