+++ doc/api/snug_logger/SnugDioLogger/onError.html
@@ -29,7 +29,7 @@
   <span id="sidenav-left-toggle" class="material-symbols-outlined" role="button" tabindex="0">menu</span>
   <ol class="breadcrumbs gt-separated dark hidden-xs">
     <li><a href="../../index.html">snug_logger</a></li>
-    <li><a href="../../snug_logger/snug_logger-library.html">snug_logger</a></li>
+    <li><a href="../../snug_logger/snug_logger-library.html">snug_logger.dart</a></li>
     <li><a href="../../snug_logger/SnugDioLogger-class.html">SnugDioLogger</a></li>
     <li class="self-crumb">onError method</li>
   </ol>
@@ -128,7 +128,7 @@ void onError(DioException err, ErrorInterceptorHandler handler) {
 <footer>
   <span class="no-break">
     snug_logger
-      1.0.4
+      1.0.8
   </span>
