+++ doc/api/snug_logger/SnugDioLogger/requestData.html
@@ -29,7 +29,7 @@
   <span id="sidenav-left-toggle" class="material-symbols-outlined" role="button" tabindex="0">menu</span>
   <ol class="breadcrumbs gt-separated dark hidden-xs">
     <li><a href="../../index.html">snug_logger</a></li>
-    <li><a href="../../snug_logger/snug_logger-library.html">snug_logger</a></li>
+    <li><a href="../../snug_logger/snug_logger-library.html">snug_logger.dart</a></li>
     <li><a href="../../snug_logger/SnugDioLogger-class.html">SnugDioLogger</a></li>
     <li class="self-crumb">requestData property</li>
   </ol>
@@ -106,7 +106,7 @@
 <footer>
   <span class="no-break">
     snug_logger
-      1.0.4
+      1.0.8
   </span>
