+++ doc/api/snug_logger/snug_logger-library-sidebar.html
@@ -2,6 +2,8 @@
       <li class="section-title"><a href="snug_logger/snug_logger-library.html#classes">Classes</a></li>
         <li><a href="snug_logger/SnugDioLogger-class.html">SnugDioLogger</a></li>
 
+      <li class="section-title"><a href="snug_logger/snug_logger-library.html#enums">Enums</a></li>
+        <li><a href="snug_logger/LogType.html">LogType</a></li>
 
 
 
@@ -9,8 +11,7 @@
       <li class="section-title"><a href="snug_logger/snug_logger-library.html#functions">Functions</a></li>
         <li><a href="snug_logger/snugLog.html">snugLog</a></li>
 
-      <li class="section-title"><a href="snug_logger/snug_logger-library.html#enums">Enums</a></li>
-        <li><a href="snug_logger/LogType.html">LogType</a></li>
+
 
 
 </ol>
