+++ doc/api/snug_logger/snug_logger-library.html
@@ -29,7 +29,7 @@
   <span id="sidenav-left-toggle" class="material-symbols-outlined" role="button" tabindex="0">menu</span>
   <ol class="breadcrumbs gt-separated dark hidden-xs">
     <li><a href="../index.html">snug_logger</a></li>
-    <li class="self-crumb">snug_logger library</li>
+    <li class="self-crumb">snug_logger.dart</li>
   </ol>
   <div class="self-name">snug_logger</div>
   <form class="search navbar-right" role="search">
@@ -56,9 +56,15 @@
       data-below-sidebar="snug_logger&#47;snug_logger-library-sidebar.html">
       
       <div>
-<h1><span class="kind-library">snug_logger</span> library 
+        
+
+        <h1>
+          <span class="kind-library">snug_logger</span>
+          library 
  
-</h1></div>
+
+        </h1>
+      </div>
 
     
 <section class="desc markdown">
@@ -86,12 +92,13 @@
 
 
 
+
     <section class="summary offset-anchor" id="functions">
       <h2>Functions</h2>
 
       <dl class="callables">
         <dt id="snugLog" class="callable">
-  <span class="name"><a href="../snug_logger/snugLog.html">snugLog</a></span><span class="signature">(<wbr><span class="parameter" id="snugLog-param-content">[<span class="type-annotation">dynamic</span> <span class="parameter-name">content</span> = <span class="default-value">&quot;Debug Message&quot;</span>, </span><span class="parameter" id="snugLog-param-logType"><span class="type-annotation"><a href="../snug_logger/LogType.html">LogType</a></span> <span class="parameter-name">logType</span> = <span class="default-value">LogType.debug</span>, </span><span class="parameter" id="snugLog-param-stackTrace"><span class="type-annotation"><a href="https://api.flutter.dev/flutter/dart-core/StackTrace-class.html">StackTrace</a>?</span> <span class="parameter-name">stackTrace</span>]</span>)
+  <span class="name"><a href="../snug_logger/snugLog.html">snugLog</a></span><span class="signature">(<wbr><span class="parameter" id="snugLog-param-content"><span class="type-annotation">dynamic</span> <span class="parameter-name">content</span>, </span><span class="parameter" id="snugLog-param-stackTrace">{<span class="type-annotation"><a href="https://api.flutter.dev/flutter/dart-core/StackTrace-class.html">StackTrace</a>?</span> <span class="parameter-name">stackTrace</span>, </span><span class="parameter" id="snugLog-param-logType"><span class="type-annotation"><a href="../snug_logger/LogType.html">LogType</a></span> <span class="parameter-name">logType</span> = <span class="default-value">LogType.debug</span>}</span>)
     <span class="returntype parameter">&#8594; dynamic</span>
   </span>
   
@@ -135,7 +142,7 @@
 
 <ol class="breadcrumbs gt-separated dark hidden-l" id="sidebar-nav">
     <li><a href="../index.html">snug_logger</a></li>
-    <li class="self-crumb">snug_logger library</li>
+    <li class="self-crumb">snug_logger.dart</li>
 </ol>
 
 
@@ -156,7 +163,7 @@
 <footer>
   <span class="no-break">
     snug_logger
-      1.0.4
+      1.0.8
   </span>
