+++ doc/api/static-assets/styles.css
@@ -164,7 +164,6 @@ main {
   flex: 1;
   display: flex;
   flex-direction: row;
-  padding: 20px;
   min-height: 0;
 }
