--- example/lib/utils/app_logger.dart
@@ -1,1 +1,0 @@
-class AppLogger {}
