+++ lib/src/base_logger/snug_logger_base.dart
@@ -3,6 +3,9 @@ import 'package:snug_logger/src/utlis/log_type.dart';
 
 /// SnugLog is a logger that provides a simple and extensible API to log messages.
 snugLog(dynamic content,
-    [LogType logType = LogType.debug, StackTrace? stackTrace]) {
-  LogHandler.snugIt(content, logType, stackTrace);
+    {StackTrace? stackTrace, LogType logType = LogType.debug}) {
+  LogHandler.snugIt(
+      content: content,
+      logType: logType,
+      stackTrace: stackTrace ?? StackTrace.current);
 }
