+++ lib/src/utlis/common_utlis.dart
@@ -5,6 +5,3 @@ class CommonUtils {
 
   static String resetColor = "\u001b[0m";
 }
-
-bool kDebugMode = !bool.fromEnvironment('dart.vm.product') &&
-    !bool.fromEnvironment('dart.vm.profile');
